+++ js/dataframe/shared/Table.svelte
@@ -991,7 +991,6 @@
 		{i18n}
 		x={active_cell_menu.x}
 		y={active_cell_menu.y}
-		col={active_cell_menu.col}
 		row={active_cell_menu?.row ?? -1}
 		on_add_row_above={() => add_row_at(active_cell_menu?.row ?? -1, "above")}
 		on_add_row_below={() => add_row_at(active_cell_menu?.row ?? -1, "below")}
@@ -1005,7 +1004,6 @@
 		{i18n}
 		x={active_header_menu.x}
 		y={active_header_menu.y}
-		col={active_header_menu.col}
 		row={-1}
 		on_add_column_left={() => add_col_at(active_header_menu?.col ?? -1, "left")}
 		on_add_column_right={() =>