+++ js/dropdown/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gradio/dropdown
 
+## 0.8.1
+
+### Dependency updates
+
+- @gradio/statustracker@0.8.1
+
 ## 0.8.0
 
 ### Features