+++ js/html/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gradio/html
 
+## 0.4.1
+
+### Dependency updates
+
+- @gradio/statustracker@0.8.1
+
 ## 0.4.0
 
 ### Features